+++ src/ReceiverLogicNew.h
@@ -12,41 +12,41 @@ private:
 	ViewModel* viewModel;
 	FlowingFFTSpectre* flowingFFTSpectre;
 
-	float spectreWidthPx = 0; //Ширина спектра в пикселях
+	double spectreWidthPx = 0; //Ширина спектра в пикселях
 
-	float centerFrequency = 0; //receiver frequency 
-	float frequencyDelta = 0; //from  -(samplerate / 2) to samplerate/2
-	float receiverPosOnPx = 0; //позиция на которой находится маркер приёма на спектре и на этой позиции ведется прослушивание станции
-	float receiverPosOnBin = 0; //позиция на которой находится маркер приёма в виде номера spectre data bin 
+	double centerFrequency = 0; //receiver frequency 
+	double frequencyDelta = 0; //from  -(samplerate / 2) to samplerate/2
+	double receiverPosOnPx = 0; //позиция на которой находится маркер приёма на спектре и на этой позиции ведется прослушивание станции
+	double receiverPosOnBin = 0; //позиция на которой находится маркер приёма в виде номера spectre data bin 
 	int savedPositionDelta = 0;
 
 public:
 
 	ReceiverLogicNew(Config* config, ViewModel* viewModel, FlowingFFTSpectre* flowingFFTSpectre);
 
-	void setFreq(float freq);
+	void setFreq(double freq);
 
-	void setCenterFrequency(float frequency);
-	float getCenterFrequency();
+	void setCenterFrequency(double frequency);
+	double getCenterFrequency();
 
-	void setFrequencyDelta(float frequencyDelta);
-	void setFrequencyDeltaBySpectrePosPx(float positionInSpectrePx);
-	void setFrequencyDeltaFromSavedPosition(float positionInSpectrePx);
-	float getFrequencyDelta();
+	void setFrequencyDelta(double frequencyDelta);
+	void setFrequencyDeltaBySpectrePosPx(double positionInSpectrePx);
+	void setFrequencyDeltaFromSavedPosition(double positionInSpectrePx);
+	double getFrequencyDelta();
 
-	void saveSpectrePositionDelta(float position);
+	void saveSpectrePositionDelta(double position);
 
-	float getPositionPX();
+	double getPositionPX();
 
-	float getSelectedFreqNew();
+	double getSelectedFreqNew();
 
-	float getPositionOnBin();
+	double getPositionOnBin();
 
-	void updateSpectreWidth(float oldSpectreWidth, float newSpectreWidth);
+	void updateSpectreWidth(double oldSpectreWidth, double newSpectreWidth);
 
-	float getFilterWidthAbs(int filterWidth);
+	double getFilterWidthAbs(int filterWidth);
 
-	float getFreqByPosOnSpectrePx(int px);
+	double getFreqByPosOnSpectrePx(int px);
 
 	void setReceivedFreqToSpectreCenter();
